requirements.txt
biopython=1.83=py38h01eb140_0
biotite=0.40.0=pypi_0
cudatoolkit=11.4.3=h39f8164_13
easydict=1.9=py_0
fair-esm=2.0.1=pypi_0
faiss-gpu=1.7.4=py3.8_hc0239a3_0_cuda11.4
filelock=3.13.3=pyhd8ed1ab_0
goatools=1.4.4=pypi_0
huggingface-hub=0.22.2=pypi_0
joblib=1.3.2=pyhd8ed1ab_0
matplotlib=3.7.3=py38h578d9bd_0
mmseqs2=15.6f452=pl5321h6a68c12_2
numpy=1.24.4=py38h59b608b_0
omegaconf=2.3.0=pypi_0
pandas=2.0.3=py38h01efb38_1
python=3.8.19=hd12c33a_0_cpython
pytorch=2.2.2=py3.8_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-scatter=2.1.2=py38_torch_2.2.0_cu121
scikit-learn=1.3.2=py38ha25d942_2
scipy=1.10.1=py38h59b608b_3
seaborn=0.12.2=py38h06a4308_0
torchmetrics=1.3.2=pypi_0
tqdm=4.66.2=pyhd8ed1ab_0
umap-learn=0.5.6=pypi_0
yaml=0.2.5=h7f98852_2

